Fix int32 overflow for large matrix sizes
For n >= 46341, n*n exceeds INT_MAX (2,147,483,647), causing signed
integer overflow in malloc size calculations and array index arithmetic.

- simple_test.c: cast m_A to size_t before multiplying in malloc calls
- NoFLA_HQRRP_WY_blk_var4.c: cast to size_t in all malloc calls with
  dimension products; replace `j + j * ldim_A` and `i + j * ldim_A`
  with `(size_t)j * ldim_A + j/i` to prevent index overflow

+++ lapack_compatible_sources/simple_test.c
@@ -53,14 +53,14 @@ int main( int argc, char *argv[] ) {
   }
 
   mn_A     = min( m_A, n_A );
-  buff_A   = ( double * ) malloc( m_A * n_A * sizeof( double ) );
+  buff_A   = ( double * ) malloc( (size_t)m_A * n_A * sizeof( double ) );
   ldim_A   = max( 1, m_A );
 
   buff_p   = ( int * ) malloc( n_A * sizeof( int ) );
 
   buff_tau = ( double * ) malloc( n_A * sizeof( double ) );
 
-  buff_Q   = ( double * ) malloc( m_A * mn_A * sizeof( double ) );
+  buff_Q   = ( double * ) malloc( (size_t)m_A * mn_A * sizeof( double ) );
   ldim_Q   = max( 1, m_A );
 
   // Generate matrix with random values in [-1, 1].
